Began cleaning up templates.H
