--- src/apcore/APGEN_Title.C
@@ -1,2 +1,0 @@
- char APGEN_build_date[] = " Fri Apr 29 06:05:00 PM PDT 2022 "; 
- char APGEN_build_id[] = "Custom Build - Fri Apr 29 18:00:19 PDT 2022";
