+++ src/apgenlib/templates.H
@@ -6,1164 +6,14 @@
 // Implementation header for the slist<> and related templates
 //
 
-#include <assert.h>
-#include <sstream>
-#include <vector>
-#include <unistd.h>
+#include "slist.H"
+#include "linkers.H"
 
-#include "treeDir.H"
-#include "C_string.H"
-#include "RES_exceptions.H"
-#include "Strategy.H"
-#include "UTL_time_base.H"
-
-
-template <class keyClass, class Node, class ownerData>
-class slist;
-
-template <class keyClass, class Node, class ownerData>
-class tlist;
+// template <class keyClass, class Node, class ownerData>
+// class tlist;
 
 template <class Tnode>
 class backptr;
-
-class safe_vector_info;
-
-//
-// NOTE on iterators and multi-list iterators ("Miterators")
-// =========================================================
-//
-// Prior to multithreading, iterators and Miterators were separate
-// classes; there was no incentive to bring them close since Use
-// Cases were different for the two.
-//
-// However, with the growing need for creating TOLs and XMLTOLs in
-// a trailing thread, it has become important to bring the two
-// classes together. The main reason for this is that evaluating
-// state resources efficiently requires turning the resource history
-// into a multi-list wich several 'threads'. A multi-resource
-// Miterator must therefore be capable of handling threads that
-// are implemented as Miterators themselves, instead of ordinary
-// iterators as was previously the case.
-//
-// As a result, the Miterator_base class, described below, has to
-// fulfill all the requirements necessary so they can be added to
-// another Miterator_base instance as just another thread.
-//
-// The key to this requirement is the Miterator::add_thread()
-// method.
-//
-
-
-//
-// A base class template for objects that iterate over multiple lists
-// ordered with respect to keys of the same type. These iterators
-// are robust against insertion and deletion of nodes from/to
-// the lists that they monitor, thanks to "handle_deletion" and
-// "handle_addition" callbacks. These callbacks are not invoked
-// by the base class templates slist and tlist below, but they
-// are invoked by the multilist template; see multilist.H.
-//
-// If you want a multi-list iterator that is robust against
-// node insertion and deletion, you therefore have to use
-// lists whose type is derived from multilist.
-//
-template <class listClass, class nodeClass>
-class Miterator_base {
-public:
-	Miterator_base()
-		: strategy(EvalTiming::NORMAL),
-			ref_count(0)
-			{}
-	virtual ~Miterator_base() {}
-
-  //
-  // The callbacks mentioned above:
-  //
-  //	callback to invoke before adding theNode:
-  //
-  virtual void		handle_addition(
-				listClass* theList,
-				nodeClass* theNode) = 0;
-
-  //
-  //	callback to invoke before removing theNode:
-  //
-  virtual void		handle_removal(
-				listClass* theList,
-				nodeClass* theNode) = 0;
-
-  //
-  // Tells the Miterator to monitor an additional list.
-  // The Miterator treats the lists it monitors as if
-  // they had been merged into one giant list, but it
-  // does not actually merge anything and is therefore
-  // much more efficient than merging the lists would
-  // be.
-  //
-  virtual void		add_thread(
-				listClass&,
-				const Cstring& name,
-				unsigned long priority,
-				bool store = false) = 0;
-
-  //
-  // Returns the number of threads that were added
-  //
-  virtual int		thread_count() = 0;
-				
-
-  //
-  // testing only
-  //
-  virtual void		test_for_refs_to(nodeClass*) = 0;
-
-  //
-  // Tells the Miterator to remove one
-  // of the lists it is monitoring
-  //
-  virtual void		remove_thread(const listClass&) = 0;
-
-  virtual const char*	get_name() const = 0;
-
-  //
-  // Removes all the monitoring data
-  //
-  virtual void		clear() = 0;
-
-  //
-  // Returns the next node in the virtual merged list
-  //
-  virtual nodeClass*	peek() = 0;
-
-  //
-  // Returns the node currently in a list owner's
-  // representative in the internal list of iterators.
-  // The void* is a pointer to the Owner member of
-  // a listClass instance. Because that pointer is not
-  // guaranteed to be non-NULL, this method is implemented
-  // trivially in simple Miterators; res_miterators
-  // have a non-trivial implementation, because resource
-  // histories are guaranteed to have an Owner that
-  // points to a valid Rsource object.
-  //
-  virtual nodeClass*	current(void*) {
-	return NULL;
-  }
-
-  //
-  // Same as peek() but updates the internal data so
-  // that the next call will return the following
-  // node
-  //
-  virtual nodeClass*	next() = 0;
-
-  //
-  // Does nothing for trailing_miter; otherwise, puts
-  // the Miterator in its initial state
-  //
-  virtual void		first() = 0;
-
-  //
-  // Returns a vector of nodeClass pointers that are
-  // safe to use by a trailing thread. This is a placeholder
-  // for all miterators, except for res_model_miter and
-  // trailing_miter.
-  //
-  virtual safe_vector_info* save_safe_nodes() {
-	return NULL;
-  }
-
-  //
-  // strategy support (evaluation of resources just before
-  // or just after a given time tag)
-  EvalTiming::Strategy	strategy;
-  
-
-  //
-  // smart_ptr support - required by the lists of threads
-  // found in derived classes
-  //
-  int			ref_count;
-
-  int			ref() {
-	ref_count++;
-	return ref_count;
-  }
-  int			deref() {
-	ref_count--;
-	return ref_count;
-  }
-
-};
-
-// nC: nodeClass
-template <class nC>
-class baseLinker {
-public:
-  baseLinker(nC* M) : me(M) {}
-  baseLinker(const baseLinker& bl) : me(bl.me) {}
-  virtual ~baseLinker() {}
-
-  nC*			me;
-  virtual void		add_cons(	TreeDir::Dir which_dir,
-					const valueHolder& countVal,
-					const valueHolder& integralVal) = 0;
-  virtual void		add_cons_to_ancestors() = 0;
-  virtual void		attach(		TreeDir::Dir right_or_left,
-					nC* node_to_attach,
-					TreeDir::cons_adjust CA) = 0;
-  virtual void		attach_left(	nC* node_to_attach,
-					TreeDir::cons_adjust CA) = 0;
-  virtual void		attach_right(	nC* node_to_attach,
-					TreeDir::cons_adjust CA) = 0;
-  virtual double	get_consumption() const = 0;
-  virtual valueHolder	get_i_cons() const = 0;
-  virtual double	get_integral() const = 0;
-  virtual double	get_integral_left() const = 0;
-  virtual double	get_integral_right() const = 0;
-  virtual true_long	get_int_consumption() const = 0;
-  virtual valueHolder	get_l_cons() const = 0;
-  virtual valueHolder	get_r_cons() const = 0;
-  virtual nC*		get_parent() const = 0;
-  virtual void		init_cons() = 0;
-  virtual bool		is_float() const { return false; }
-  virtual void		set_i_cons(const valueHolder* vh) = 0;
-  virtual void		set_parent(nC*) = 0;
-  virtual void		subtract_cons_from_ancestors(nC* stop_here = NULL) = 0;
-};
-
-template <class nC>
-class simpleLinker: public baseLinker<nC> {
-public:
-
-  simpleLinker(nC* bn) : baseLinker<nC>(bn) {}
-  simpleLinker(const simpleLinker& sl) : baseLinker<nC>(sl) {}
-  ~simpleLinker() {}
-
-  void			add_cons(	TreeDir::Dir which_dir,
-					const valueHolder& countVal,
-					const valueHolder& integralVal) {}
-  void			add_cons_to_ancestors() {}
-  void			attach(		TreeDir::Dir right_or_left,
-					nC* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_left(	nC* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_right(	nC* node_to_attach,
-					TreeDir::cons_adjust CA);
-  double		get_consumption() const { return 0.0; }
-  true_long		get_int_consumption() const { return 0; }
-  valueHolder		get_i_cons() const { return valueHolder(0L); }
-  double		get_integral() const { return 0.0; }
-  double		get_integral_left() const { return 0.0; }
-  double		get_integral_right() const { return 0.0; }
-  valueHolder		get_l_cons() const { return valueHolder(0L); }
-  valueHolder		get_r_cons() const { return valueHolder(0L); }
-  nC*			get_parent() const { return NULL; }
-  void			init_cons() {}
-  void			set_i_cons(const valueHolder* vh) {}
-  void			set_parent(nC*) {}
-  void			subtract_cons_from_ancestors(nC* stop_here = NULL) {} };
-
-template <class bN>
-void
-simpleLinker<bN>::attach_left(bN* N, TreeDir::cons_adjust CA) {
-	attach(TreeDir::LEFT_DIR, N, TreeDir::UPDATE_CONS); }
-
-template <class bN>
-void
-simpleLinker<bN>::attach_right(bN* N, TreeDir::cons_adjust CA) {
-	attach(TreeDir::RIGHT_DIR, N, TreeDir::UPDATE_CONS); }
-
-template <class bN>
-void
-simpleLinker<bN>::attach(TreeDir::Dir right_or_left, bN* N, TreeDir::cons_adjust CA) {
-	baseLinker<bN>::me->Links[right_or_left] = N; }
-
-
-template <class bN>
-class intLinker: public baseLinker<bN> {
-public:
-  bN*			avlParent;
-  valueHolder		myCount;
-  valueHolder		nCount_right;
-  valueHolder		nCount_left;
-
-  intLinker(bN* bn) :	baseLinker<bN>(bn),
-			avlParent(NULL),
-			myCount(1L),
-			nCount_right(0L),
-			nCount_left(0L) {}
-  intLinker(const intLinker& il) : baseLinker<bN>(il) ,
-			avlParent(NULL),
-			myCount(1L),
-			nCount_right(0L),
-			nCount_left(0L) {}
-  ~intLinker() {}
-
-  void			add_cons(	TreeDir::Dir which_dir,
-					const valueHolder& countVal,
-					const valueHolder& integralVal);
-  void			add_cons_to_ancestors();
-  void			attach(		TreeDir::Dir right_or_left,
-					bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_left(	bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_right(	bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  double		get_consumption() const { return (double) get_int_consumption(); }
-  true_long		get_int_consumption() const;
-  valueHolder		get_i_cons() const { return myCount; }
-  double		get_integral() const { return 0.0; }
-  double		get_integral_left() const { return 0.0; }
-  double		get_integral_right() const { return 0.0; }
-  valueHolder		get_l_cons() const { return nCount_left; }
-  valueHolder		get_r_cons() const { return nCount_right; }
-  bN*			get_parent() const { return avlParent; }
-  virtual void		init_cons() {
-				avlParent = NULL;
-				nCount_left = valueHolder(0L);
-				nCount_right = valueHolder(0L); }
-  void			set_i_cons(const valueHolder* vh) { myCount = vh ? *vh : valueHolder(0L); }
-  void			set_parent(bN* b) { avlParent = b; }
-  void			subtract_cons_from_ancestors(bN* stop_here = NULL); };
-
-template <class bN>
-void
-intLinker<bN>::add_cons(
-		TreeDir::Dir which_dir,
-		const valueHolder& the_amount,
-		const valueHolder& /* integralVal */) {
-	if(which_dir == TreeDir::RIGHT_DIR) {
-		nCount_right += the_amount; }
-	else {
-		nCount_left += the_amount; } }
-
-template <class bN>
-void
-intLinker<bN>::add_cons_to_ancestors() {
-	bN* p = baseLinker<bN>::me->get_parent();
-	bN* t = baseLinker<bN>::me;
-
-	while(p) {
-		if(t == p->Links[TreeDir::LEFT_DIR]) {
-			p->add_cons(TreeDir::LEFT_DIR, myCount, 0.0); }
-		else {
-			p->add_cons(TreeDir::RIGHT_DIR, myCount, 0.0); }
-		t = p;
-		p = p->get_parent(); } }
-
-template <class bN>
-void
-intLinker<bN>::subtract_cons_from_ancestors(
-		bN* until_this) {
-	bN* p = baseLinker<bN>::me->get_parent();
-	bN* t = baseLinker<bN>::me;
-
-	while(p) {
-		if(t == p->Links[TreeDir::LEFT_DIR]) {
-			p->add_cons(TreeDir::LEFT_DIR, valueHolder(-myCount.get_content()), 0L); }
-		else {
-			p->add_cons(TreeDir::RIGHT_DIR, valueHolder(-myCount.get_content()), 0L); }
-		t = p;
-		p = p->get_parent();
-		if(p == until_this) {
-			break; } } }
-
-template <class bN>
-void
-intLinker<bN>::attach_left(
-		bN* N,
-		TreeDir::cons_adjust CA) {
-	attach(TreeDir::LEFT_DIR, N, CA); }
-
-template <class bN>
-void
-intLinker<bN>::attach_right(
-		bN* N,
-		TreeDir::cons_adjust CA) {
-	attach(TreeDir::RIGHT_DIR, N, CA); }
-
-template <class bN>
-void
-intLinker<bN>::attach(
-		TreeDir::Dir right_or_left,
-		bN* N,
-		TreeDir::cons_adjust CA) {
-	bN* an = baseLinker<bN>::me;
-	an->Links[right_or_left] = N;
-	if(N) {
-		N->set_parent(baseLinker<bN>::me);
-		if(CA == TreeDir::UPDATE_CONS) {
-			valueHolder	t(N->get_i_cons());
-			t += N->get_l_cons();
-			t += N->get_r_cons();
-			if(right_or_left == TreeDir::RIGHT_DIR) {
-				nCount_right = t; }
-			else {
-				nCount_left = t; } } }
-	else if(CA == TreeDir::UPDATE_CONS) {
-		if(right_or_left == TreeDir::RIGHT_DIR) {
-			nCount_right = valueHolder(0L); }
-		else {
-			nCount_left = valueHolder(0L); } } }
-
-template <class bN>
-true_long
-intLinker<bN>::get_int_consumption() const {
-	true_long val = get_i_cons().get_content() + get_l_cons().get_content();
-	const bN* q = baseLinker<bN>::me;
-	const bN* p = avlParent;
-	while(p) {
-		if(q == p->get_link(TreeDir::RIGHT_DIR)) {
-			val += p->get_i_cons().get_content() + p->get_l_cons().get_content(); }
-		q = p;
-		p = p->get_parent(); }
-	return val; }
-
-template <class bN>
-class floatLinker: public baseLinker<bN> {
-public:
-  bN*			avlParent;
-  valueHolder		myCount;
-  valueHolder		nCount_right;
-  valueHolder		nCount_left;
-  valueHolder		integral_right, integral_left, my_integral;
-
-  floatLinker(bN* bn) : baseLinker<bN>(bn),
-			avlParent(NULL), myCount(0.0), nCount_right(0.0),
-			nCount_left(0.0), my_integral(0.0),
-			integral_right(0.0), integral_left(0.0) {}
-  floatLinker(const floatLinker& fl) : baseLinker<bN>(fl),
-			avlParent(NULL), myCount(0.0), nCount_right(0.0),
-			nCount_left(0.0), my_integral(0.0),
-			integral_right(0.0), integral_left(0.0) {}
-  ~floatLinker() {}
-
-  void			add_cons(	TreeDir::Dir which_dir,
-					const valueHolder& countVal,
-					const valueHolder& integralVal);
-  void			add_cons_to_ancestors();
-  void			attach(		TreeDir::Dir right_or_left,
-					bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_left(	bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  void			attach_right(	bN* node_to_attach,
-					TreeDir::cons_adjust CA);
-  double		get_consumption() const;
-  true_long		get_int_consumption() const { return 0; }
-  double		get_integral() const;
-  double		get_integral_left() const { return integral_left.get_double(); }
-  double		get_integral_right() const { return integral_right.get_double(); }
-  true_long		get_int_consumption() { return 0; }
-  valueHolder		get_i_cons() const { return myCount; }
-  valueHolder		get_l_cons() const { return nCount_left; }
-  valueHolder		get_r_cons() const { return nCount_right; }
-  bN*			get_parent() const { return avlParent; }
-  virtual void		init_cons() {
-				avlParent = NULL;
-				nCount_left = valueHolder(0.0);
-				nCount_right = valueHolder(0.0);
-				// ??
-				// myCount = valueHolder(0.0);
-				my_integral = 0.0;
-				integral_right = 0.0;
-				integral_left = 0.0; }
-  bool			is_float() const { return true; }
-  void			set_i_cons(const valueHolder* vh) { myCount = vh ? *vh : valueHolder(0.0); }
-  void			set_parent(bN* b) { avlParent = b; }
-  void			subtract_cons_from_ancestors(bN* stop_here = NULL); };
-
-template <class bN>
-void
-floatLinker<bN>::add_cons(
-		TreeDir::Dir which_dir,
-		const valueHolder &V,
-		const valueHolder &I) {
-	if(which_dir == TreeDir::RIGHT_DIR) {
-		integral_right += I;
-		nCount_right += V; }
-	else {
-		integral_left += I;
-		nCount_left += V; } }
-
-template <class bN>
-void
-floatLinker<bN>::add_cons_to_ancestors() {
-	bN*		p = baseLinker<bN>::me->get_parent();
-	bN*		t = baseLinker<bN>::me;
-	CTime_base	t0(baseLinker<bN>::me->getKey().getetime());
-	double		myTime = t0.convert_to_double_use_with_caution();
-	double		deltaT;
-
-	while(p) {
-		CTime_base	t1(p->getKey().getetime());
-		deltaT = t1.convert_to_double_use_with_caution() - myTime;
-		if(t == p->Links[TreeDir::LEFT_DIR]) {
-			p->add_cons(TreeDir::LEFT_DIR, myCount, valueHolder(myCount.get_double() * deltaT)); }
-		else {
-			p->add_cons(TreeDir::RIGHT_DIR, myCount, valueHolder(myCount.get_double() * deltaT)); }
-		t = p;
-		p = p->get_parent(); } }
-
-template <class bN>
-double
-floatLinker<bN>::get_consumption() const {
-	double val = get_i_cons().get_double() + get_l_cons().get_double();
-	const bN* q = baseLinker<bN>::me;
-	const bN* p = avlParent;
-	while(p) {
-		if(q == p->Links[TreeDir::RIGHT_DIR]) {
-			val += p->get_i_cons().get_double() + p->get_l_cons().get_double(); }
-		q = p;
-		p = p->get_parent(); }
-	return val; }
-
-template <class bN>
-double
-floatLinker<bN>::get_integral() const {
-	double val = integral_left.get_double();
-	bN* q = baseLinker<bN>::me;
-	bN* p = avlParent;
-	CTime_base t1 = baseLinker<bN>::me->getKey().getetime(), t2;
-	double deltaT;
-
-	while(p) {
-		if(q == p->Links[TreeDir::RIGHT_DIR]) {
-			t2 = p->getKey().getetime();
-			deltaT = t1.convert_to_double_use_with_caution() - t2.convert_to_double_use_with_caution();
-			val += deltaT * (p->get_i_cons().get_double() + p->get_l_cons().get_double())
-				+ p->get_integral_left(); }
-		q = p;
-		p = p->get_parent(); }
-	return val; }
-
-// keyClass'd better be alpha_time...
-template <class bN>
-void
-floatLinker<bN>::subtract_cons_from_ancestors(
-		bN* until_this) {
-	bN*	p = get_parent();
-	bN*	t = baseLinker<bN>::me;
-	CTime_base t0 = baseLinker<bN>::me->getKey().getetime();
-	double	deltaT;
-	double	myTime = t0.convert_to_double_use_with_caution();
-
-	while(p) {
-		CTime_base t1(p->getKey().getetime());
-		deltaT = t1.convert_to_double_use_with_caution() - myTime;
-		if(t == p->Links[TreeDir::LEFT_DIR]) {
-			p->add_cons(TreeDir::LEFT_DIR, valueHolder(-myCount.get_double()), valueHolder(-deltaT * myCount.get_double())); }
-		else {
-			p->add_cons(TreeDir::RIGHT_DIR, valueHolder(-myCount.get_double()), valueHolder(-deltaT * myCount.get_double())); }
-		t = p;
-		p = p->get_parent();
-		if(p == until_this) {
-			break; } } }
-
-template <class bN>
-void
-floatLinker<bN>::attach(
-		TreeDir::Dir right_or_left,
-		bN* af,
-		TreeDir::cons_adjust CA) {
-	bN* an = baseLinker<bN>::me;
-	an->Links[right_or_left] = af;
-	if(af) {
-		af->set_parent(baseLinker<bN>::me);
-		if(CA == TreeDir::UPDATE_CONS) {
-			valueHolder		t(af->get_i_cons());
-			CTime_base		t1 = baseLinker<bN>::me->getKey().getetime(),
-						t2 = af->getKey().getetime();
-			double			deltaT;
-			double			u = af->get_integral_left() + af->get_integral_right();
-
-			deltaT = t1.convert_to_double_use_with_caution() - t2.convert_to_double_use_with_caution();
-			t += af->get_l_cons();
-			t += af->get_r_cons();
-			u += t.get_double() * deltaT;
-			if(right_or_left == TreeDir::RIGHT_DIR) {
-				integral_right = u;
-				nCount_right = t; }
-			else {
-				integral_left = u;
-				nCount_left = t; } } }
-	else if(CA == TreeDir::UPDATE_CONS) {
-		if(right_or_left == TreeDir::RIGHT_DIR) {
-			integral_right = valueHolder(0.0);
-			nCount_right = valueHolder(0.0); }
-		else {
-			integral_left = valueHolder(0.0);
-			nCount_left = valueHolder(0.0); } } }
-
-template <class bN>
-void
-floatLinker<bN>::attach_left(
-		bN* N,
-		TreeDir::cons_adjust CA) {
-	attach(TreeDir::LEFT_DIR, N, CA); }
-
-template <class bN>
-void
-floatLinker<bN>::attach_right(
-		bN* N,
-		TreeDir::cons_adjust CA) {
-	attach(TreeDir::RIGHT_DIR, N, CA); }
-
-template <class keyClass, class nbase, class ownerData = short>
-class slist {
-public:
-
-private:
-  nbase*	head;
-  nbase*	tail;
-  long		length;
-
-public:
-  ownerData	Owner;
-
-  slist();
-  slist(const ownerData&);
-  slist(const slist &);
-  virtual ~slist();
-
-  class iterator {
-    public:
-
-    //
-    // DATA
-    //
-    const slist*	L;
-    nbase*		iterNode;
-
-    //
-    // CONSTRUCTORS
-    //
-    iterator() = delete;
-    iterator(const slist& l) {
-	L = &l;
-	iterNode = NULL;
-    }
-    iterator(const slist* l) {
-	L = l;
-	iterNode = NULL;
-    }
-    iterator(const iterator& li) {
-	L = li.L;
-	iterNode = li.iterNode;
-    }
-    virtual ~iterator() {}
-
-    //
-    // METHODS
-    //
-    virtual nbase*		getNode() const {
-	return iterNode;
-    }
-    virtual nbase*		getPrev() const {
-	return NULL;
-    }
-    virtual bool		is_empty() const {
-	return iterNode == NULL;
-    }
-    virtual const void*	get_void_list() const {
-	return (void*)L;
-    }
-    virtual iterator& operator=(const iterator& li) {
-	L = li.L;
-	iterNode = li.iterNode;
-	return *this;
-    }
-    virtual nbase*	operator()() {
-	return next();
-    }
-    virtual nbase*	const_first() const {
-	return L->first_node();
-    }
-
-    //
-    // this method is trivial for ordinary iterators; it isn't for
-    // derived classes such as flexIterators.
-    //
-    virtual nbase*	first() {
-	go_to_beginning();
-	return iterNode;
-    }
-    virtual void	go_to_beginning() {
-	iterNode = L->first_node();
-    }
-    virtual void	go_to_end() {
-	iterNode = L->last_node();
-    }
-    virtual void	go_to(nbase* N) {
-	iterNode = N;
-    }
-    virtual nbase*	last() {
-	go_to_end();
-	return iterNode;
-    }
-    virtual nbase*	next() {
-	if(iterNode) {
-	    if(!iterNode->next_node()) {
-		return (iterNode = NULL);
-	    } else {
-	    	return (iterNode = iterNode->next_node());
-	    }
-	} else {
-
-	    //
-	    // sets iterNode to the first node of the list:
-	    //
-	    first();
-	}
-	return iterNode;
-    }
-
-    //
-    // Provides the caller with a glimpse of what next()
-    // will return, but does not change the content of
-    // the iterator:
-    //
-    virtual nbase*	peek() const {
-	return iterNode ?
-	    iterNode->next_node()
-	    : NULL;
-    }
-    virtual nbase*	previous() {
-	if(iterNode) {
-	    iterNode = iterNode->previous_node();
-	} else {
-	    iterNode = L->last_node();
-	}
-	return iterNode;
-    }
-    virtual void	undefine() {
-	iterNode = NULL;
-    }
-  }; // end of iterator subclass
-
-//
-// 		INVARIANTS for curval_iter
-// 		--------------------------
-//
-//	See the README.md document in the src directory.
-//
-  class curval_iter: public iterator {
-    public:
-    //
-    // DATA
-    //
-    nbase*		prevNode;
-
-    //
-    // CONSTRUCTORS
-    //
-    curval_iter() = delete;
-    curval_iter(const slist& l)
-    	    : iterator(l) {
-	prevNode = NULL;
-    }
-    curval_iter(const slist* l)
-    	    : iterator(l) {
-	prevNode = NULL;
-    }
-    curval_iter(const curval_iter& li)
-    	    : iterator(li) {
-	prevNode = li.prevNode;
-    }
-    virtual		~curval_iter() {}
-
-    //
-    // METHODS
-    //
-    virtual nbase*	getPrev() const override {
-	return prevNode;
-    }
-    virtual curval_iter& operator=(const curval_iter& li) {
-	iterator::L = li.L;
-	iterator::iterNode = li.iterNode;
-	prevNode = li.prevNode;
-	return *this;
-    }
-    virtual void	go_to_beginning() override {
-	iterator::iterNode = iterator::L->first_node();
-
-	//
-	// Not totally logical, but serves the purpose of resource
-	// histories
-	//
-	prevNode = iterator::iterNode;
-    }
-    virtual void	go_to_end() override {
-	iterator::iterNode = iterator::L->last_node();
-	if(iterator::iterNode) {
-		prevNode = iterator::iterNode->previous_node();
-	} else {
-		prevNode = NULL;
-	}
-    }
-    virtual void	go_to(nbase* N) override {
-	iterator::iterNode = N;
-	if(N) {
-		prevNode = iterator::iterNode->previous_node();
-	} else {
-		prevNode = NULL;
-	}
-    }
-    virtual nbase*	next() override {
-	if(iterator::iterNode) {
-	    if(!iterator::iterNode->next_node()) {
-
-		prevNode = iterator::iterNode;
-		iterator::iterNode = NULL;
-		return prevNode;
-	    }
-
-	    //
-	    // Special considerations for resource history nodes
-	    // =================================================
-	    //
-	    // When the iterator reaches the end of history, the iterNode
-	    // returned will be NULL but the prevNode pointer points to
-	    // the last event in the history list. This allows currentval()
-	    // to work just the way it should.
-	    //
-	    // When potential_triggers() advances through the history list
-	    // during the second scheduling pass, prevNode should always
-	    // point to a "2nd generation" node, i. e., a node that is
-	    // created during the scheduling pass proper, and not left
-	    // over from the preceding modeling (or scheduling) pass.
-	    //
-	    prevNode = iterator::iterNode;
-	    return (iterator::iterNode = iterator::iterNode->next_node());
-	} else {
-	    iterator::first();
-	}
-	return iterator::iterNode;
-    }
-    virtual nbase*	previous() override {
-	if(iterator::iterNode) {
-	    iterator::iterNode = iterator::iterNode->previous_node();
-	} else {
-	    iterator::iterNode = iterator::L->last_node();
-	}
-	if(iterator::iterNode) {
-	    prevNode = iterator::iterNode->previous_node();
-	}
-	return iterator::iterNode;
-    }
-    virtual void		undefine() override {
-	iterator::iterNode = NULL;
-	prevNode = NULL;
-    }
-  }; // end of curval_iter subclass
-
-//
-// 		INVARIANTS for leading_iter
-// 		---------------------------
-//
-// 	- safeNode has an event_id that is strictly
-// 	  greater than mEvent::lastEventIDbeforeScheduling,
-// 	  and will therefore never be deleted by the event
-// 	  loop as long as it is active.
-//
-  class leading_iter : public iterator {
-    public:
-
-    //
-    // DATA
-    //
-
-
-    //
-    // Too slow:
-    //
-    // std::atomic<nbase*> safeNode;
-
-    //
-    // safeNode
-    nbase*	safeNode;
-
-    //
-    // CONSTRUCTORS
-    //
-    leading_iter() = delete;
-    leading_iter(const slist& l)
-    	    : iterator(l) {
-	safeNode = NULL;
-    }
-    leading_iter(const slist* l)
-    	    : iterator(l) {
-	safeNode = NULL;
-    }
-    leading_iter(const leading_iter& li)
-   	    : iterator(li) {
-	safeNode = li.safeNode;
-    }
-    virtual		~leading_iter() {}
-
-    //
-    // METHODS
-    //
-    virtual nbase*	getPrev() const override {
-	return safeNode;
-    }
-    virtual leading_iter& operator=(const leading_iter& li) {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-	return *this;
-    }
-    virtual void	go_to_beginning() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-    }
-    virtual void	go_to_end() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-	iterator::iterNode = iterator::L->last_node();
-	if(iterator::iterNode) {
-		safeNode = iterator::iterNode->previous_node();
-	} else {
-		safeNode = NULL;
-	}
-    }
-    virtual void	go_to(nbase* N) override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-    }
-    virtual nbase*	last() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-	return iterator::iterNode;
-    }
-    virtual nbase*	next() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-	return iterator::iterNode;
-    }
-    virtual nbase*	previous() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-    }
-    virtual void	undefine() override {
-
-	//
-	// We want the model_res_miter to have complete control
-	//
-	assert(false);
-    }
-  }; // end of leading_iter subclass
-
-//
-// The following class requires a pointer to a leading_iter so it can
-// navigate the list pointed to by the iterator. To do so, one has to
-// follow what the modeling thread is doing, hence have access to the
-// safeNode member of a leading_iter. The data member pointing to this
-// leading_iter is called leader.
-//
-// Here are the invariants for this class:
-//
-// 			INVARIANTS
-// 			----------
-//
-// 	- leader->safeNode has a time tag that is less than
-// 	  or equal to the modeling thread's value of 'now'
-//
-// 	- leader->safeNode has an event_id that is strictly
-// 	  greater than mEvent::lastEventIDbeforeScheduling,
-// 	  and will therefore never be deleted by the event
-// 	  loop as long as it is active
-//
-// As a result, if a node N was initialized to a safe node,
-// and if the time tag of N is strictly less than 'now',
-// and N is not equal to the current value of leader->safeNode,
-// then it is safe to request N->next_node().
-//
-// The reason is that the modeling thread will only update
-// leader->safeNode to nodes in the list which have no predecessor
-// created by the previous modeling or scheduling pass. We know
-// this, because the modeling thread only updates safeNode to
-// nodes with a time tag less than or equal to now, and by the
-// time nodes are being created with now having a certain value,
-// old node processing has eliminated all the nodes with a time
-// tag less than or equal to now. Therefore, it is safe to
-// navigate the list of safe nodes as long as one keeps to the
-// part of the list that is at or before leader->safeNode.
-//
-  class safe_iter: public curval_iter {
-  public:
-
-	safe_iter() = delete;
-	safe_iter(const safe_iter&) = delete;
-	safe_iter(const slist& L, nbase* I)
-	    : curval_iter(L),
-		safe_node(I) {
-	    iterator::iterNode = I;
-	    curval_iter::prevNode = I;
-	}
-	~safe_iter() {}
-
-	//
-	// There are no methods here; all interesting
-	// methods are found in the trailing_miter class,
-	// which manages safe_iter directly.
-	//
-
-	//
-	// the lone data member:
-	//
-	nbase*	safe_node;
-
-  }; // end of safe_iter subclass
-
-  //
-  // The following methods are provided for the benefit of lists
-  // that are "under observation" by Miterators.
-  //
-  // The idea is that whenever the list is modified through the
-  // addition or removal of nodes), it updates any Miterators
-  // that are watching it so that the Miterators can update their
-  // state. The purpose of these state updates is to make sure
-  // that (1) no pointers are pointing to nodes that are about
-  // to be deleted, and (2) if a node is inserted just before
-  // the "next node" in an iterator, the iterator is modified
-  // to point to the new node instead of the one it is currently
-  // pointing to.
-  //
-  virtual void		notify_iterators_about_adding(nbase*) {}
-  virtual void		notify_iterators_about_removing(nbase*) {}
-
-  //
-  // to help nodes that may be inserted in a tlist:
-  //
-  virtual void		make_root(nbase*) {}
-
-  //
-  // testing only
-  //
-  virtual void		test_iterators_for_refs_to(nbase*) {}
-
-  //
-  // The rationale for including the last two arguments is that
-  // they are useful when the Miterator being registered is one
-  // that should be used by a resource when asked to evaluate
-  // its current value. Here is how a typical scenario plays out:
-  //
-  // 1. a new std::thread defines a miterator - for example,
-  //    a trailing_miter, which it intends to use when evaluating
-  //    the current (for that thread) value of some resources.
-  //    Let's denote this miterator by M.
-  //
-  // 2. for each resource the std::thread is interested in, it
-  //    adds M to the resource's history list, H.
-  //
-  // 3. H knows which resource owns it, thanks to the Owner
-  //    template class argument. H is in fact a multilist,
-  //	which overrides register_miterator(). In the overriding
-  //	method, the resource is identified and its
-  //	push_miterator_data() method is invoked, with the
-  //	miterator and the iterator as arguments.
-  //
-  // 4. In its evaluate_present_value() method, the resource
-  //    uses the miterator data as follows: the miterator
-  //    provides the evaluation strategy, and the iterator
-  //    provides the pointer to the history node that is
-  //    the closest to the time at which currentval()
-  //    needs to be evaluated.
-  //
-  // UPDATE
-  // ======
-  //
-  // To enable concurrent modeling and TOL creation, it is
-  // necessary to simplify the structure of state histories.
-  // This means breaking up the single history list into
-  // separate lists, one for set/reset events and one for
-  // _each_ use event. As a result, the registration process
-  // has become more complicated.
-  //
-  // Remember that part of the registration functionality is
-  // to provide the resource that owns a list as part of its
-  // history with means to evaluate its current value by
-  // only querying the currently "active node" in its history.
-  // The active node used to be an ordinary node belonging
-  // to a well-defined list. But now, the active node belongs
-  // to whatever 'use' list is in a Miterator's active threads
-  // list; if there is no active 'use' list, then the active
-  // node is the current node in the set/reset list, which
-  // resembles the existing situation.
-  //
-  virtual void		register_miterator(
-				Miterator_base<slist, nbase>&,
-				iterator* = NULL,
-				bool = false) {}
-  virtual void		unregister_miterator(
-				Miterator_base<slist, nbase>&) {}
-  virtual void		unregister_miterators() {}
-  virtual int		iterator_count() const {
-	return 0;
-  }
-
-  virtual void		check() { return; }
-  virtual bool		check_consumption(ostringstream& s) {
-	return true;
-  }
-  virtual void		clear();
-  bool			delete_node(nbase* aNode);
-  virtual nbase*	find(const keyClass& s) const;
-
-			//
-			// In slist, this is the same as find(s); do not
-			// use... the purpose is to have a virtual base for
-			// the tlist version, which actually does what it says
-			//
-  virtual nbase*	find_at_or_before(const keyClass& s) const {
-				Cstring err("Do not use slist<...>::find_at_or_before()");
-				throw(eval_error(err));
-			}
-
-			/* 'after' means 'strictly after'. Key match must be
-			 * exact, else return NULL. */
-  virtual nbase*	find_match_after(const keyClass& s, nbase*) const;
-  virtual nbase*	find_adjacent_right(const nbase*) const { return NULL; }
-  virtual nbase*	find_adjacent_left(const nbase*) const { return NULL; }
-  virtual nbase*	first_node() const { return head; }
-  long			get_length() const { return length; }
-  nbase*		insert_first_node_after_second(nbase* insert_this, nbase* after_that);
-  virtual nbase*	insert_node(nbase* N);
-  virtual nbase*	last_node() const { return tail; }
-  slist&		operator = (const slist &);
-  virtual slist&	operator << (nbase* N) { insert_node(N); return *this; }
-  virtual slist&	operator << (slist& Arg);
-  
-  virtual nbase*	remove_node(nbase* aNode);
-  virtual nbase*	remove_node_prior_to_deletion(nbase* N) {
-	  return remove_node(N);
-  }
-};	// end of class slist
 
 template <class keyClass, class nbase, class ownerData = short>
 class tlist: public slist<keyClass, nbase, ownerData> {
@@ -1225,11 +75,6 @@ public:
   tlist&		operator << (lbase&) override;
   tlist&		operator << (tlist &);
   bool			operator <= (tlist &);
-
-#ifdef TRY_WITHOUT
-			// returns nonsense unless nLinker = intLinker
-  virtual nbase*	operator[] (const unsigned int i) const;
-#endif /* TRY_WITHOUT */
 
 }; // end of class tlist
